refactor(util): move the bundle capability check into its own module
First cut of the split #293 asks for. `dependencies.py` owns three
separable concerns and had grown to 2248 lines; this takes the one that
answers "can the bundled maidr.js draw what we are about to hand it".

The seam is one the tests already drew: `tests/core/test_bundle_capability.py`
and `tests/core/test_bundle_freshness.py` split along exactly this line
before the module did.

`_RELEASE_RE` deliberately stays behind despite sitting in the moved
region -- `_version_key` uses it, and it is a version-parsing primitive
rather than a capability one. `_bundle_warning_enabled` and
`BUNDLE_WARNING_ENV_VAR` stay too, and are imported: the warning policy is
shared with the freshness side, so it belongs to neither alone.

`maidr.util.dependencies` keeps resolving the moved names through a lazy
module `__getattr__` rather than a re-export, because `bundle_capability`
imports *from* it and an eager re-export would make the cycle real.

Refs #293
