+++ maidr/plotly/plotly_maidr.py
@@ -35,6 +35,10 @@ from maidr.plotly.step_shape import (
     is_step_trace,
     renders_through_webgl,
 )
+from maidr.util.bundle_capability import (
+    schema_trace_types,
+    warn_if_bundle_cannot_render,
+)
 from maidr.util.dependencies import (
     MAIDR_JS_FILENAME,
     OFFLINE_FALLBACK_REPORT,
@@ -44,8 +48,6 @@ from maidr.util.dependencies import (
     maidr_bundled_relative_dir,
     maidr_html_dependency,
     maidr_js_cdn_url,
-    schema_trace_types,
-    warn_if_bundle_cannot_render,
     warn_if_bundle_is_stale,
 )
 from maidr.util.environment import Environment
